+++ docs/api/changelog.rst
@@ -13,6 +13,7 @@ Changelog
     * Removed ``hp.Nope``
     * Removed a number of other helpers that are too random for this library
     * Removing ``hp.ResultStreamer`` to be replaced with ``hp.QueueFeeder``
+    * Renamed ``hp.ATicker`` to ``hp.Ticker``
     * Machinery is now Python3.13+
 
 .. _release-0.1.1: